+++ packages/graph-explorer/src/components/Tabular/useTabular.ts
@@ -432,7 +432,6 @@ export const useTabular = <T extends object>(options: TabularOptions<T>) => {
       autoResetSortBy: !skipPageResetRef.current,
       autoResetFilters: !skipPageResetRef.current,
       autoResetRowState: !skipPageResetRef.current,
-      /* eslint-enable react-hooks/refs */
     },
     useColumnOrder,
     useResizeColumns,